begin getting ready for package
