+++ lib/lottie.dart
@@ -1,0 +1,2 @@
+export 'package:lottie_flutter/src/lottie.dart';
+export 'package:lottie_flutter/src/composition.dart';
